+++ app/controllers/puffer/grid_base.rb
@@ -1,0 +1,11 @@
+module Puffer
+  class GridBase < Puffer::Base
+    unloadable
+
+    setup do
+      per_page 60
+    end
+
+    #helper 'puffer/helpers/puffer_grid'
+  end
+end