+++ spec/lib/filters_spec.rb
@@ -5,7 +5,7 @@ describe Puffer::Filters do
   describe 'controller filters class generation' do
 
     it 'should set proper attributes' do
-      Puffer::Filters.new(Admin::PostsController.filter_fields).attributes.keys.should == %w(user.email status title body puffer_search puffer_order)
+      Puffer::Filters.new(Admin::PostsController.filter_fields).attributes.keys.should == %w(user.email status title body puffer_search puffer_order puffer_per_page)
     end
 
     it 'should set proper query' do